+++ app/src/main/java/com/rittmann/myapplication/main/draw/DrawObject.kt
@@ -3,7 +3,7 @@ package com.rittmann.myapplication.main.draw
 import android.graphics.Canvas
 
 interface DrawObject {
-    fun update(deltaTime: Float)
+    fun update(deltaTime: Double)
     fun draw(canvas: Canvas)
     fun free()
 }
