+++ app/src/main/java/com/rittmann/myapplication/main/entity/Pointer.kt
@@ -21,7 +21,7 @@ class Pointer(
         paint.color = Color.YELLOW
     }
 
-    override fun update(deltaTime: Float) {}
+    override fun update(deltaTime: Double) {}
 
     override fun draw(canvas: Canvas) {
         canvas.drawRect(rect, paint)
